refactor DataClient and MessageClient
